+++ Library/Homebrew/bundle/package_type.rb
@@ -1,0 +1,6 @@
+# typed: strict
+# frozen_string_literal: true
+
+# TODO: Remove this compatibility require once bundle callers load
+# `bundle/extensions/extension` directly.
+require "bundle/extensions/extension"
