+++ Library/Homebrew/bundle/package_types.rb
@@ -1,0 +1,8 @@
+# typed: strict
+# frozen_string_literal: true
+
+require "bundle/package_type"
+require "bundle/tap"
+require "bundle/brew"
+require "bundle/cask"
+require "bundle/extensions"
